% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdnify_rmd.R
\name{cdnify_rmd}
\alias{cdnify_rmd}
\title{CDNify an .Rmd HTML Document}
\usage{
cdnify_rmd(filename, libdir = sub(".html$", "_files", filename))
}
\arguments{
\item{filename}{.html file to CDNify.}

\item{libdir}{Directory containing dependencies (defaults to filename_files).}
}
\description{
This function replaces dependencies in HTML Documents knit from
.Rmd files.
}
\details{
The page must have been knit with \code{self_contained: false} in the
YAML header. For example, a valid header would be:

\preformatted{---
title: "An Example RMarkdown Document"
author: "Engelbert Humperdinck"
output:
  html_document:
    self_contained: false
---}
}
