% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdnify_tufte.R
\name{cdnify_tufte}
\alias{cdnify_tufte}
\title{CDNify a \strong{tufte} page}
\usage{
cdnify_tufte(filename, libdir = sub(".html$", "_files", filename))
}
\arguments{
\item{filename}{.html file to CDNify.}

\item{libdir}{Directory containing dependencies (defaults to filename_files).}
}
\description{
This function replaces dependencies in .html pages generated using
the \href{https://github.com/rstudio/tufte}{tufte} package.
}
\details{
The page must have been knit with \code{self_contained: false} in the
YAML header. For example, a valid header would be:

\preformatted{---
title: "An Example Using the Tufte Style"
author: "John Smith"
output:
  tufte::tufte_html:
    tufte_features: ["fonts", "background", "italics"]
    self_contained: false
---}
}
